requirements.txt
streamlit
pandas
scikit-learn
joblib
SQLAlchemy
PyMySQL
python-dotenv
matplotlib
openpyxl
xlrd
numpy